requirements.txt
psycopg2>=2.9.1
pyyaml>=6.0
requests>=2.20.0